Add files via upload
This code implements a simple JavaScript-based number guessing game. The game generates a random number between 1 and 20 (inclusive) and the user must guess what that number is.

Functionalities:

The game starts by displaying the message "Start guessing..." and the score as 20.
The user can make a guess by entering a number in the input field and clicking on the "Check" button.
If the user's guess is incorrect, the game will display either "Too High 📈" or "Too Low 📉", and the score will be decremented by 1.
If the user's score reaches 0, the game will display the message "You Lost The Game 😒" and the score will be set to 0.
If the user's guess is correct, the game will display the message "Correct Number 🎉" and the secret number will be revealed. The background color of the page will change to green, and the high score will be updated if the current score is higher than the previous high score.
The user can play again by clicking on the "Again" button, which will reset the score, secret number, message, and input field.